Fix for PWM out of sync issue during clip-capture
Neuroptics reported that the PWM list sent to MCU is not in sync with
the clip captures frames received.

The list send by them should turn PWM off for 15 frames and ON for next
15 frames. Instead only the first 13 frames were OFF and the PWM turns
ON from 14th frame(It should be ON from the 16th frame). So the reason
for issue is  PWM synchronization starts at two frames earlier.

By investigating the 3 dummy frames which are skipped at beginning of clip
capture, the PWM list starts earlier at the 2nd dummy frame.

Fix:
By raising the sync GPIO to turn on two frames later the issue can be
fixed.